tests/wpt/tests/css/css-gaps/animation/gap-decorations-outset-neutral-keyframe-002.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>gap decorations column-rule-outset properties support neutral keyframe</title>
    <link rel="help" href="https://drafts.csswg.org/css-gaps-1/#outset">
    <meta name="assert" content="gap decorations column-rule-outset value list supports neutral keyframe.">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/css/support/interpolation-testcommon.js"></script>
  </head>
  <body>
    <div id="target"></div>
    <script>
      test(() => {
        target.style.columnRuleStyle = 'solid';
        target.style.columnRuleEdgeStartOutset = '1px';
        target.style.columnRuleEdgeEndOutset = '1px';
        target.style.columnRuleInteriorStartOutset = '1px';
        target.style.columnRuleInteriorEndOutset = '1px';

        var animation = target.animate([{}, {columnRuleEdgeStartOutset: '11px'}], 1000);
        animation.pause();
        animation.currentTime = 500;
        assert_equals(getComputedStyle(target).columnRuleEdgeStartOutset, '6px');

        var animation2 = target.animate([{}, {columnRuleEdgeEndOutset: '11px'}], 1000);
        animation2.pause();
        animation2.currentTime = 500;
        assert_equals(getComputedStyle(target).columnRuleEdgeEndOutset, '6px');

        var animation3 = target.animate([{}, {columnRuleInteriorStartOutset: '11px'}], 1000);
        animation3.pause();
        animation3.currentTime = 500;
        assert_equals(getComputedStyle(target).columnRuleInteriorStartOutset, '6px');

        var animation4 = target.animate([{}, {columnRuleInteriorEndOutset: '11px'}], 1000);
        animation4.pause();
        animation4.currentTime = 500;
        assert_equals(getComputedStyle(target).columnRuleInteriorEndOutset, '6px');
      }, 'gap decorations column-rule-outset properties support neutral keyframe.');
    </script>
  </body>
</html>
